Fix the generation of Swift code for defined type aliases.
In a declaration like "Foo": { "type": "number", "format": "double" }, where
there is not a "properties" entry in that declaration, the type Foo is
simply an alias for Double.  These were previously being generated in Swift
as an empty class, but then they wouldn't decode because on the wire there
would be a number, not a dictionary.

Fix this by generating a typealias declaration in this situation.